--- a/src/cases/camelcase/mod.rs
+++ b/src/cases/camelcase/mod.rs
@@ -1,13 +1,8 @@
 use std::ascii::*;
-use regex::Regex;
-
-use cases::classcase::is_class_case;
-use cases::snakecase::to_snake_case;
-
 /// Converts a `String` to camelCase `String`
 ///
 /// #Examples
-/// ```
+/// ```0
 ///     use inflector::cases::camelcase::to_camel_case;
 ///     let mock_string: String = "fooBar".to_string();
 ///     let expected_string: String = "fooBar".to_string();
@@ -15,7 +10,7 @@ use cases::snakecase::to_snake_case;
 ///     assert!(asserted_string == expected_string);
 ///
 /// ```
-/// ```
+/// ```1
 ///     use inflector::cases::camelcase::to_camel_case;
 ///     let mock_string: String = "FOO_BAR".to_string();
 ///     let expected_string: String = "fooBar".to_string();
@@ -23,7 +18,7 @@ use cases::snakecase::to_snake_case;
 ///     assert!(asserted_string == expected_string);
 ///
 /// ```
-/// ```
+/// ```2
 ///     use inflector::cases::camelcase::to_camel_case;
 ///     let mock_string: String = "Foo Bar".to_string();
 ///     let expected_string: String = "fooBar".to_string();
@@ -31,7 +26,7 @@ use cases::snakecase::to_snake_case;
 ///     assert!(asserted_string == expected_string);
 ///
 /// ```
-/// ```
+/// ```3
 ///     use inflector::cases::camelcase::to_camel_case;
 ///     let mock_string: String = "foo_bar".to_string();
 ///     let expected_string: String = "fooBar".to_string();
@@ -39,7 +34,7 @@ use cases::snakecase::to_snake_case;
 ///     assert!(asserted_string == expected_string);
 ///
 /// ```
-/// ```
+/// ```4
 ///     use inflector::cases::camelcase::to_camel_case;
 ///     let mock_string: String = "Foo bar".to_string();
 ///     let expected_string: String = "fooBar".to_string();
@@ -47,7 +42,7 @@ use cases::snakecase::to_snake_case;
 ///     assert!(asserted_string == expected_string);
 ///
 /// ```
-/// ```
+/// ```5
 ///     use inflector::cases::camelcase::to_camel_case;
 ///     let mock_string: String = "foo-bar".to_string();
 ///     let expected_string: String = "fooBar".to_string();
@@ -55,7 +50,7 @@ use cases::snakecase::to_snake_case;
 ///     assert!(asserted_string == expected_string);
 ///
 /// ```
-/// ```
+/// ```6
 ///     use inflector::cases::camelcase::to_camel_case;
 ///     let mock_string: String = "FooBar".to_string();
 ///     let expected_string: String = "fooBar".to_string();
@@ -63,7 +58,7 @@ use cases::snakecase::to_snake_case;
 ///     assert!(asserted_string == expected_string);
 ///
 /// ```
-/// ```
+/// ```7
 ///     use inflector::cases::camelcase::to_camel_case;
 ///     let mock_string: String = "FooBar3".to_string();
 ///     let expected_string: String = "fooBar3".to_string();
@@ -72,30 +67,31 @@ use cases::snakecase::to_snake_case;
 ///
 /// ```
 pub fn to_camel_case(non_camelized_string: String) -> String {
-    to_camel_from_snake(to_snake_case(non_camelized_string))
-}
-
-fn to_camel_from_snake(non_camelized_string: String) -> String {
-    let mut result: String = "".to_string();
     let mut new_word: bool = false;
-
-    for character in non_camelized_string.chars() {
-        if character.to_string() == "_" {
-            new_word = true;
-        } else if new_word {
-            result = format!("{}{}", result, character.to_ascii_uppercase());
-            new_word = false;
-        } else {
-            result = format!("{}{}", result, character.to_ascii_lowercase());
-        }
-    }
-    result
+    let mut last_char: char = ' ';
+    non_camelized_string
+        .chars()
+        .fold("".to_string(), |result, character|
+            if character == '-' || character == '_' || character == ' ' {
+                new_word = true;
+                result
+            } else if new_word || (
+                (last_char.is_lowercase() && character.is_uppercase()) &&
+                (last_char != ' ')
+                ){
+                new_word = false;
+                format!("{}{}", result, character.to_ascii_uppercase())
+            } else {
+                last_char = character;
+                format!("{}{}", result, character.to_ascii_lowercase())
+            }
+        )
 }
 
 /// Determines if a `String` is camelCase bool``
 ///
 /// #Examples
-/// ```
+/// ```0
 ///     use inflector::cases::camelcase::is_camel_case;
 ///     let mock_string: String = "Foo".to_string();
 ///     let asserted_bool: bool = is_camel_case(mock_string);
@@ -103,15 +99,15 @@ fn to_camel_from_snake(non_camelized_string: String) -> String {
 ///
 ///
 /// ```
-/// ```
+/// ```1
 ///     use inflector::cases::camelcase::is_camel_case;
 ///     let mock_string: String = "foo".to_string();
 ///     let asserted_bool: bool = is_camel_case(mock_string);
-///     assert!(asserted_bool == false);
+///     assert!(asserted_bool == true);
 ///
 ///
 /// ```
-/// ```
+/// ```2
 ///     use inflector::cases::camelcase::is_camel_case;
 ///     let mock_string: String = "foo-bar-string-that-is-really-really-long".to_string();
 ///     let asserted_bool: bool = is_camel_case(mock_string);
@@ -119,7 +115,7 @@ fn to_camel_from_snake(non_camelized_string: String) -> String {
 ///
 ///
 /// ```
-/// ```
+/// ```3
 ///     use inflector::cases::camelcase::is_camel_case;
 ///     let mock_string: String = "FooBarIsAReallyReallyLongString".to_string();
 ///     let asserted_bool: bool = is_camel_case(mock_string);
@@ -127,7 +123,7 @@ fn to_camel_from_snake(non_camelized_string: String) -> String {
 ///
 ///
 /// ```
-/// ```
+/// ```4
 ///     use inflector::cases::camelcase::is_camel_case;
 ///     let mock_string: String = "fooBarIsAReallyReally3LongString".to_string();
 ///     let asserted_bool: bool = is_camel_case(mock_string);
@@ -135,7 +131,7 @@ fn to_camel_from_snake(non_camelized_string: String) -> String {
 ///
 ///
 /// ```
-/// ```
+/// ```5
 ///     use inflector::cases::camelcase::is_camel_case;
 ///     let mock_string: String = "fooBarIsAReallyReallyLongString".to_string();
 ///     let asserted_bool: bool = is_camel_case(mock_string);
@@ -143,7 +139,7 @@ fn to_camel_from_snake(non_camelized_string: String) -> String {
 ///
 ///
 /// ```
-/// ```
+/// ```6
 ///     use inflector::cases::camelcase::is_camel_case;
 ///     let mock_string: String = "FOO_BAR_STRING_THAT_IS_REALLY_REALLY_LONG".to_string();
 ///     let asserted_bool: bool = is_camel_case(mock_string);
@@ -151,7 +147,7 @@ fn to_camel_from_snake(non_camelized_string: String) -> String {
 ///
 ///
 /// ```
-/// ```
+/// ```7
 ///     use inflector::cases::camelcase::is_camel_case;
 ///     let mock_string: String = "foo_bar_string_that_is_really_really_long".to_string();
 ///     let asserted_bool: bool = is_camel_case(mock_string);
@@ -159,7 +155,7 @@ fn to_camel_from_snake(non_camelized_string: String) -> String {
 ///
 ///
 /// ```
-/// ```
+/// ```8
 ///     use inflector::cases::camelcase::is_camel_case;
 ///     let mock_string: String = "Foo bar string that is really really long".to_string();
 ///     let asserted_bool: bool = is_camel_case(mock_string);
@@ -167,20 +163,14 @@ fn to_camel_from_snake(non_camelized_string: String) -> String {
 ///
 ///
 /// ```
-/// ```
+/// ```9
 ///     use inflector::cases::camelcase::is_camel_case;
 ///     let mock_string: String = "Foo Bar Is A Really Really Long String".to_string();
 ///     let asserted_bool: bool = is_camel_case(mock_string);
 ///     assert!(asserted_bool == false);
 /// ```
 pub fn is_camel_case(test_string: String) -> bool {
-    let camel_matcher = Regex::new(r"(^|[A-Z])([^-|^_|^ ]*[a-z0-9]+[A-Z][a-z0-9]+)").unwrap();
-    let kebab_snake_matcher = Regex::new(r"[-|_| ]").unwrap();
-    if camel_matcher.is_match(test_string.as_ref()) &&
-       !kebab_snake_matcher.is_match(test_string.as_ref()) && !is_class_case(test_string) {
-        return true;
-    }
-    false
+    to_camel_case(test_string.clone()) == test_string
 }
 
 #[cfg(test)]
@@ -189,17 +179,22 @@ mod tests {
     use self::test::Bencher;
 
     #[bench]
-    fn bench_camel(b: &mut Bencher) {
+    fn bench_camel0(b: &mut Bencher) {
         b.iter(|| super::to_camel_case("Foo bar".to_string()));
+    }
+
+    #[bench]
+    fn bench_camel1(b: &mut Bencher) {
+        b.iter(|| super::to_camel_case("foo_bar".to_string()));
+    }
+
+    #[bench]
+    fn bench_camel2(b: &mut Bencher) {
+        b.iter(|| super::to_camel_case("fooBar".to_string()));
     }
 
     #[bench]
     fn bench_is_camel(b: &mut Bencher) {
         b.iter(|| super::is_camel_case("Foo bar".to_string()));
-    }
-
-    #[bench]
-    fn bench_camel_from_sname(b: &mut Bencher) {
-        b.iter(|| super::to_camel_from_snake("foo_bar".to_string()));
     }
 }