--- a/src/cases/classcase/mod.rs
+++ b/src/cases/classcase/mod.rs
@@ -208,18 +208,18 @@ mod tests {
     extern crate test;
     use self::test::Bencher;
 
-    #[bench]
-    fn bench_class_case(b: &mut Bencher) {
-        b.iter(|| super::to_class_case("Foo bar".to_string()));
-    }
-
-    #[bench]
-    fn bench_is_class(b: &mut Bencher) {
-        b.iter(|| super::is_class_case("Foo bar".to_string()));
-    }
-
-    #[bench]
-    fn bench_class_from_snake(b: &mut Bencher) {
-        b.iter(|| super::to_class_from_snake("foo_bar".to_string()));
-    }
+    // #[bench]
+    // fn bench_class_case(b: &mut Bencher) {
+    //     b.iter(|| super::to_class_case("Foo bar".to_string()));
+    // }
+    //
+    // #[bench]
+    // fn bench_is_class(b: &mut Bencher) {
+    //     b.iter(|| super::is_class_case("Foo bar".to_string()));
+    // }
+    //
+    // #[bench]
+    // fn bench_class_from_snake(b: &mut Bencher) {
+    //     b.iter(|| super::to_class_from_snake("foo_bar".to_string()));
+    // }
 }