--- a/src/cases/mod.rs
+++ b/src/cases/mod.rs
@@ -47,3 +47,5 @@ pub mod uppercase;
 ///
 /// Example string `LOWERCASE`
 pub mod lowercase;
+
+mod case;