… Cargo.toml
… README.md
… changelog.md
… src/cases/sentencecase/mod.rs
… src/cases/snakecase/mod.rs
… src/cases/tablecase/mod.rs
… src/cases/titlecase/mod.rs
… src/cases/uppercase/mod.rs
… src/numbers/deordinalize/mod.rs
… src/numbers/mod.rs
… src/string/mod.rs
… src/string/singularize/mod.rs
… src/suffix/mod.rs