@NullVoxPopuli 2016-09-26
is this faster than the implementation done by @calebmer https://github.com/whatisinternet/inflector/issues/18#issuecomment-213879759
?