@whatisinternet 2016-09-27
@NullVoxPopuli I only benched against camel case on his crate but here are the results:

```shell
// Inflections
test case::tests::bench_camel0   ... bench:         424 ns/iter (+/- 89)
test case::tests::bench_camel1   ... bench:         441 ns/iter (+/- 134)
test case::tests::bench_camel2   ... bench:         412 ns/iter (+/- 167)
test case::tests::bench_is_camel ... bench:         445 ns/iter (+/- 198)
```

```shell
// Inflector
test cases::camelcase::tests::bench_camel0                      ... bench:         139 ns/iter (+/- 40)
test cases::camelcase::tests::bench_camel1                      ... bench:         138 ns/iter (+/- 31)
test cases::camelcase::tests::bench_camel2                      ... bench:         138 ns/iter (+/- 41)
test cases::camelcase::tests::bench_is_camel                    ... bench:         184 ns/iter (+/- 90)
```

looks like Inflector is faster now 👍. I'm happy to bench the rest of his crate tomorrow. If you're looking at using `Inflector` in your project I'd be interested in the benchmarks on your side too.