@whatisinternet 2016-09-27
# Complete bench results from [inflections](https://github.com/calebmer/inflections):

**Note**: Inflections is missing some of the features of inflector thus only the features supported are benchmarked)

**Note**: All benchmarks were run using `rustc 1.13.0-nightly (4f9812a59 2016-09-21)`

```shell
test case::tests::bench_camel0              ... bench:         425 ns/iter (+/- 91)
test case::tests::bench_camel1              ... bench:         434 ns/iter (+/- 169)
test case::tests::bench_camel2              ... bench:         403 ns/iter (+/- 184)
test case::tests::bench_is_camel            ... bench:         464 ns/iter (+/- 79)
test case::tests::bench_is_kebab            ... bench:         336 ns/iter (+/- 145)
test case::tests::bench_is_lower            ... bench:         284 ns/iter (+/- 102)
test case::tests::bench_is_screaming_snake  ... bench:         470 ns/iter (+/- 612)
test case::tests::bench_is_sentence         ... bench:         343 ns/iter (+/- 89)
test case::tests::bench_is_snake            ... bench:         347 ns/iter (+/- 178)
test case::tests::bench_is_title            ... bench:         425 ns/iter (+/- 113)
test case::tests::bench_is_upper            ... bench:         302 ns/iter (+/- 50)
test case::tests::bench_kebab               ... bench:         317 ns/iter (+/- 94)
test case::tests::bench_kebab_from_snake    ... bench:         660 ns/iter (+/- 224)
test case::tests::bench_lower               ... bench:         287 ns/iter (+/- 127)
test case::tests::bench_screaming_snake     ... bench:         315 ns/iter (+/- 106)
test case::tests::bench_sentence            ... bench:         333 ns/iter (+/- 130)
test case::tests::bench_sentence_from_snake ... bench:         334 ns/iter (+/- 74)
test case::tests::bench_snake_from_camel    ... bench:         319 ns/iter (+/- 110)
test case::tests::bench_snake_from_snake    ... bench:         712 ns/iter (+/- 300)
test case::tests::bench_snake_from_title    ... bench:         333 ns/iter (+/- 232)
test case::tests::bench_title               ... bench:         438 ns/iter (+/- 199)
test case::tests::bench_title_from_snake    ... bench:         465 ns/iter (+/- 155)
test case::tests::bench_upper               ... bench:         306 ns/iter (+/- 47)
```

# Benchmark code used on `inflections`:
**Note**: Run on `master` as of posting
```rust
#[cfg(test)]
mod tests {
    extern crate test;
    use self::test::Bencher;

    #[bench]
    fn bench_camel0(b: &mut Bencher) {
        b.iter(|| super::to_camel_case("Foo bar"));
    }

    #[bench]
    fn bench_camel1(b: &mut Bencher) {
        b.iter(|| super::to_camel_case("foo_bar"));
    }

    #[bench]
    fn bench_camel2(b: &mut Bencher) {
        b.iter(|| super::to_camel_case("fooBar"));
    }

    #[bench]
    fn bench_is_camel(b: &mut Bencher) {
        b.iter(|| super::is_camel_case("Foo bar"));
    }
    #[bench]
    fn bench_kebab(b: &mut Bencher) {
        b.iter(|| super::to_kebab_case("Foo bar"));
    }

    #[bench]
    fn bench_is_kebab(b: &mut Bencher) {
        b.iter(|| super::is_kebab_case("Foo bar"));
    }

    #[bench]
    fn bench_kebab_from_snake(b: &mut Bencher) {
        b.iter(|| super::to_kebab_case("test_test_test"));
    }
    #[bench]
    fn bench_lower(b: &mut Bencher) {
        b.iter(|| super::to_lower_case("Foo BAR"));
    }

    #[bench]
    fn bench_is_lower(b: &mut Bencher) {
        b.iter(|| super::is_lower_case("Foo bar"));
    }
    #[bench]
    fn bench_screaming_snake(b: &mut Bencher) {
        b.iter(|| super::to_constant_case("Foo bar"));
    }

    #[bench]
    fn bench_is_screaming_snake(b: &mut Bencher) {
        b.iter(|| super::is_constant_case("Foo bar"));
    }
    #[bench]
    fn bench_sentence(b: &mut Bencher) {
        b.iter(|| super::to_sentence_case("Foo BAR"));
    }

    #[bench]
    fn bench_is_sentence(b: &mut Bencher) {
        b.iter(|| super::is_sentence_case("Foo bar"));
    }

    #[bench]
    fn bench_sentence_from_snake(b: &mut Bencher) {
        b.iter(|| super::to_sentence_case("foo_bar"));
    }
    #[bench]
    fn bench_snake_from_title(b: &mut Bencher) {
        b.iter(|| super::to_snake_case("Foo bar"));
    }

    #[bench]
    fn bench_snake_from_camel(b: &mut Bencher) {
        b.iter(|| super::to_snake_case("fooBar"));
    }

    #[bench]
    fn bench_snake_from_snake(b: &mut Bencher) {
        b.iter(|| super::to_snake_case("foo_bar_bar_bar"));
    }

    #[bench]
    fn bench_is_snake(b: &mut Bencher) {
        b.iter(|| super::is_snake_case("Foo bar"));
    }
    #[bench]
    fn bench_title(b: &mut Bencher) {
        b.iter(|| super::to_title_case("Foo BAR"));
    }

    #[bench]
    fn bench_is_title(b: &mut Bencher) {
        b.iter(|| super::is_title_case("Foo bar"));
    }

    #[bench]
    fn bench_title_from_snake(b: &mut Bencher) {
        b.iter(|| super::to_title_case("foo_bar"));
    }
    #[bench]
    fn bench_upper(b: &mut Bencher) {
        b.iter(|| super::to_upper_case("Foo BAR"));
    }

    #[bench]
    fn bench_is_upper(b: &mut Bencher) {
        b.iter(|| super::is_upper_case("Foo bar"));
    }
}

```